PMM-6863 Renames dirs.
