+++ k8s_api/pxc/v1/pxc_backup_types.go
@@ -17,7 +17,7 @@
 package v1
 
 import (
-	"github.com/percona-platform/dbaas-controller/k8_api/meta"
+	"github.com/percona-platform/dbaas-controller/k8s_api/meta"
 )
 
 // PerconaXtraDBClusterBackupList holds exported fields representing Percona XtraDB cluster backup list.
